event schema moved to event.py
